Refactor VisNetwork enums
