+++ HtmlForgeX/HtmlForgeX.csproj
@@ -81,6 +81,7 @@
 
     <ItemGroup>
         <Folder Include="Containers\ScrollingText\" />
+        <Folder Include="Containers\VisNetwork\Enums\" />
     </ItemGroup>
 
     <!-- <ItemGroup>
